[refactor][s]: Rename script to be more specific
  * Update imports.
