+++ tests/test_ckan_to_frictionless_mapping.py
@@ -1,12 +1,12 @@
 import json
 
-import ckan_datapackage_tools
+import ckan_datapackage_tools.ckan_to_frictionless as ckan_to_frictionless
 
 
 def test_resource():
     inpath = 'tests/fixtures/ckan_resource.json'
     exppath = 'tests/fixtures/frictionless_resource.json'
-    converter = ckan_datapackage_tools.CKANToFrictionless()
+    converter = ckan_to_frictionless.CKANToFrictionless()
     indict = json.load(open(inpath))
     exp = json.load(open(exppath))
     out = converter.resource(indict)
@@ -16,7 +16,7 @@ def test_resource():
 def test_package():
     inpath = 'tests/fixtures/ckan_package.json'
     exppath = 'tests/fixtures/frictionless_package.json'
-    converter = ckan_datapackage_tools.CKANToFrictionless()
+    converter = ckan_to_frictionless.CKANToFrictionless()
     indict = json.load(open(inpath))
     exp = json.load(open(exppath))
     out = converter.package(indict)
